Move function to transformations helpers
